domain: add reference counting to domain_info
This is not a functional change, as there currently is only one holder
of the reference.

Holding an active reference on a domain implicitly means holding a ref
on its info object. However, this will allow us to pin info objects
even without holding an active reference, which is needed for
references that should outlive a single syscall. This will also allow
info objects to outlive their handles, even though that is not
something we would want.

Signed-off-by: Tom Gundersen <teg@jklm.no>